requirements.txt
django
djangorestframework
python-dotenv
bcrypt
PyJWT
psycopg2-binary